# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.OportunidadeMEI:
    properties:
      bairro:
        example: Centro
        type: string
      cidade:
        example: Rio de Janeiro
        type: string
      cnae_ids:
        example:
        - "8130300"
        - "8121400"
        items:
          type: string
        type: array
      complemento:
        example: Próximo ao Paço Imperial
        type: string
      cover_image:
        description: Imagens
        example: https://example.com/imagem.jpg
        type: string
      created_at:
        example: "2024-01-01T10:00:00Z"
        type: string
      data_expiracao:
        example: "2024-12-31T23:59:59Z"
        type: string
      data_limite_execucao:
        example: "2024-12-31T23:59:59Z"
        type: string
      deleted_at:
        example: "2024-12-31T23:59:59Z"
        type: string
      descricao_servico:
        example: Serviço de poda, jardinagem e manutenção das áreas verdes
        type: string
      estado:
        example: RJ
        type: string
      forma_pagamento:
        description: Pagamento e prazos
        enum:
        - CHEQUE
        - DINHEIRO
        - CARTAO
        - PIX
        - TRANSFERENCIA
        - ""
        example: PIX
        type: string
      gallery_images:
        example:
        - https://example.com/img1.jpg
        - https://example.com/img2.jpg
        items:
          type: string
        type: array
      id:
        example: 1
        type: integer
      logradouro:
        description: Local de execução
        example: Praça XV de Novembro
        type: string
      numero:
        example: S/N
        type: string
      orgao_id:
        description: Relacionamentos
        example: SECONSERVA
        type: string
      outras_informacoes:
        example: Trabalho a ser realizado nos finais de semana
        type: string
      prazo_pagamento:
        example: 30 dias após conclusão
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.StatusOportunidadeMEI'
        description: Status e controle
        enum:
        - draft
        - active
        - expired
        example: active
      titulo:
        example: Manutenção de Jardins da Praça XV
        type: string
      updated_at:
        example: "2024-01-01T10:00:00Z"
        type: string
    type: object
...
  models.StatusOportunidadeMEI:
    enum:
    - draft
    - active
    - expired
    type: string
    x-enum-varnames:
    - StatusOportunidadeDraft
    - StatusOportunidadeActive
    - StatusOportunidadeExpired
...
host: localhost:8081
info:
  contact:
    email: frederico.zolio@prefeitura.rio
    name: API Support
    url: https://iplan.rio
  description: API de serviços para aplicativos da Prefeitura do Rio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Go
  version: "1.0"
paths:
...
  /api/public/oportunidades-mei:
    get:
      description: Retorna uma lista paginada de oportunidades MEI ativas
      parameters:
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (default: 10, max: 1000)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por órgão
        in: query
        name: orgaoId
        type: string
      - description: Filtrar por status (draft, active, expired)
        in: query
        name: status
        type: string
      - description: Buscar por título (case-insensitive)
        in: query
        name: titulo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar oportunidades MEI
      tags:
      - oportunidades-mei
  /api/public/oportunidades-mei/{id}:
    get:
      description: Retorna uma oportunidade MEI pelo seu ID
      parameters:
      - description: ID da oportunidade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Obter oportunidade MEI por ID
      tags:
      - oportunidades-mei
...
  /api/v1/oportunidades-mei:
    get:
      description: Retorna uma lista paginada de oportunidades MEI ativas
      parameters:
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (default: 10, max: 1000)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por órgão
        in: query
        name: orgaoId
        type: string
      - description: Filtrar por status (draft, active, expired)
        in: query
        name: status
        type: string
      - description: Buscar por título (case-insensitive)
        in: query
        name: titulo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar oportunidades MEI
      tags:
      - oportunidades-mei
    post:
      consumes:
      - application/json
      description: Cria uma nova oportunidade MEI publicada
      parameters:
      - description: Dados da oportunidade
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OportunidadeMEI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Criar oportunidade MEI
      tags:
      - oportunidades-mei
  /api/v1/oportunidades-mei/{id}:
    delete:
      description: Remove uma oportunidade MEI pelo ID (soft delete)
      parameters:
      - description: ID da oportunidade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Excluir oportunidade MEI
      tags:
      - oportunidades-mei
    get:
      description: Retorna uma oportunidade MEI pelo seu ID
      parameters:
      - description: ID da oportunidade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Obter oportunidade MEI por ID
      tags:
      - oportunidades-mei
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma oportunidade MEI existente mantendo seu
        status atual
      parameters:
      - description: ID da oportunidade
        in: path
        name: id
        required: true
        type: integer
      - description: Dados da oportunidade
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OportunidadeMEI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Atualizar oportunidade MEI
      tags:
      - oportunidades-mei
...
  /api/v1/oportunidades-mei/{id}/publish:
    put:
      description: Publica uma oportunidade MEI que estava em rascunho
      parameters:
      - description: ID da oportunidade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Publicar oportunidade MEI
      tags:
      - oportunidades-mei
  /api/v1/oportunidades-mei/draft:
    post:
      consumes:
      - application/json
      description: Cria uma nova oportunidade MEI em status de rascunho
      parameters:
      - description: Dados da oportunidade
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OportunidadeMEI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OportunidadeMEI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Criar oportunidade MEI como rascunho
      tags:
      - oportunidades-mei
  /api/v1/oportunidades-mei/drafts:
    get:
      description: Retorna uma lista paginada de oportunidades MEI em rascunho
      parameters:
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (default: 10, max: 1000)'
        in: query
        name: pageSize
        type: integer
      - description: Filtrar por órgão
        in: query
        name: orgaoId
        type: string
      - description: Buscar por título (case-insensitive)
        in: query
        name: titulo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listar rascunhos de oportunidades MEI
      tags:
      - oportunidades-mei
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Autenticação usando token JWT no formato 'Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"